requirements.txt
tensorflow-gpu==1.10.0
keras==2.3.1
numpy==1.14.2
python-dotenv==0.13.0
nltk==3.2.5
flask==1.1.2